refactor: signup inputs
