+++ foodway/app/src/main/java/com/example/foodway/model/Input.kt
@@ -1,0 +1,9 @@
+package com.example.foodway.model
+
+import androidx.annotation.DrawableRes
+
+data class Input(
+    val inputLabel: String,
+    @DrawableRes
+    val icon: Int
+)
