+++ foodway/app/src/main/java/com/example/foodway/view/components/InputGeneric.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.OutlinedTextFieldDefaults
@@ -23,7 +22,7 @@ import androidx.compose.ui.unit.dp
 import com.example.foodway.R
 
 @Composable
-fun Input(inputLabel: String, icon: Int) {
+fun InputGeneric(inputLabel: String, icon: Int) {
     var label by remember { mutableStateOf("") }
 
     OutlinedTextField(
